backend/pyproject.toml
[project]
name = "traceback-backend"
version = "0.1.0"
description = "TraceBack - AI多智能体因果回溯分析引擎"
requires-python = ">=3.11"
license = { text = "GPL-2.0" }
authors = [
    { name = "TraceBack Team" }
]

dependencies = [
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    "openai>=1.0.0",
    "duckduckgo-search>=7.0.0",
    "requests>=2.31.0",
    "PyMuPDF>=1.24.0",
    "python-docx>=1.0.0",
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "zep-cloud>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
